cutalyst-backend/requirements.txt
fastapi
uvicorn
python-multipart
cloudinary
openai-whisper
ffmpeg-python
